Task 7. Refactoring structure
