+++ main/src/main/java/org/example/pages/sqlEx/SqlExMainPage.java
@@ -1,12 +1,12 @@
 package org.example.pages.sqlEx;
 
 import io.qameta.allure.Step;
-import org.example.pages.MainPage;
+import org.example.pages.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class SqlExMainPage extends MainPage {
+public class SqlExMainPage extends BasePage {
 
     private static final String URL = "https://www.sql-ex.ru";
