+++ main/src/main/java/org/example/pages/way2automation/practice/bankingApp/BankManagerLoginPage.java
@@ -1,7 +1,7 @@
 package org.example.pages.way2automation.practice.bankingApp;
 
 import io.qameta.allure.Step;
-import org.example.pages.MainPage;
+import org.example.pages.base.BasePage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -11,7 +11,7 @@ import org.openqa.selenium.support.ui.Select;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-public class BankManagerLoginPage extends MainPage {
+public class BankManagerLoginPage extends BasePage {
 
     public static final String URL = "https://www.way2automation.com/angularjs-protractor/banking/#/manager";
