+++ main/src/main/java/org/example/pages/way2automation/practice/way2autoJquery/FramesAndWindowsJqueryPage.java
@@ -1,0 +1,4 @@
+package org.example.pages.way2automation.practice.way2autoJquery;
+
+public class FramesAndWindowsJqueryPage {
+}
