+++ main/src/main/java/org/example/pages/way2automation/publicSite/Way2AutomationMainPage.java
@@ -1,14 +1,14 @@
 package org.example.pages.way2automation.publicSite;
 
 import io.qameta.allure.Step;
-import org.example.pages.MainPage;
+import org.example.pages.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
 import java.util.List;
 
-public class Way2AutomationMainPage extends MainPage {
+public class Way2AutomationMainPage extends BasePage {
 
     private static final String URL = "https://www.way2automation.com/";
