+++ main/src/main/java/org/example/pages/wiki/WikipediaMainPage.java
@@ -1,12 +1,12 @@
 package org.example.pages.wiki;
 
 import io.qameta.allure.Step;
-import org.example.pages.MainPage;
+import org.example.pages.base.BasePage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class WikipediaMainPage extends MainPage {
+public class WikipediaMainPage extends BasePage {
 
     private static final String URL = "https://ru.wikipedia.org/wiki/";
